+++ app/users/dashboard/page.tsx
@@ -9,16 +9,21 @@ import {
   CardDescription,
   CardHeader,
   CardTitle,
-} from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
+} from '@/components/shadcn/card';
+import { Button } from '@/components/shadcn/button';
 
 import { toast } from '@/lib/styles/toast-styles';
-import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
+import {
+  Tabs,
+  TabsContent,
+  TabsList,
+  TabsTrigger,
+} from '@/components/shadcn/tabs';
 import {
   ChartContainer,
   ChartTooltip,
   ChartTooltipContent,
-} from '@/components/ui/chart';
+} from '@/components/shadcn/chart';
 import {
   Bar,
   BarChart,
